+++ examples/tls/tls_common.h
@@ -65,7 +65,7 @@
 #endif
 
 /* enable benchmarking mode */
-#if 0
+#ifndef NO_TPM_BENCH
     #define TLS_BENCH_MODE
 #endif
 
@@ -314,22 +314,6 @@ static inline void CloseAndCleanupSocket(SockIoCbCtx* sockIoCtx)
 /******************************************************************************/
 /* --- BEGIN Supporting TLS functions --- */
 /******************************************************************************/
-
-#ifdef TLS_BENCH_MODE
-    #include <sys/time.h>
-    static inline double gettime_secs(int reset)
-    {
-    #ifdef XTIME
-    	extern double current_time(int reset);
-        return current_time(reset);
-    #else
-        struct timeval tv;
-        gettimeofday(&tv, 0);
-        (void)reset;
-        return (double)tv.tv_sec + (double)tv.tv_usec / 1000000;
-    #endif
-    }
-#endif
 
 static inline int myVerify(int preverify, WOLFSSL_X509_STORE_CTX* store)
 {
