+++ src/tpm2_wrap.c
@@ -276,7 +276,7 @@ int wolfTPM2_Cleanup(WOLFTPM2_DEV* dev)
         return BAD_FUNC_ARG;
     }
 
-#if defined(WOLF_CRYPTO_DEV) || defined(WOLF_CRYPTO_CB)
+#if !defined(WOLFTPM2_NO_WOLFCRYPT) && (defined(WOLF_CRYPTO_DEV) || defined(WOLF_CRYPTO_CB))
     /* make sure crypto dev callback is unregistered */
     rc = wolfTPM2_ClearCryptoDevCb(dev, INVALID_DEVID);
     if (rc != 0)
@@ -2242,6 +2242,7 @@ static int wolfTPM2_ComputeSymmetricUnique(WOLFTPM2_DEV* dev, int hashAlg,
         wc_HashFree(&hash, hashType);
     }
 #else
+    (void)hashAlg;
     rc = NOT_COMPILED_IN;
 #endif
 
@@ -2814,7 +2815,7 @@ int wolfTPM2_GetNvAttributesTemplate(TPM_HANDLE auth, word32* nvAttributes)
 /******************************************************************************/
 
 
-#if defined(WOLF_CRYPTO_DEV) || defined(WOLF_CRYPTO_CB)
+#if !defined(WOLFTPM2_NO_WOLFCRYPT) && (defined(WOLF_CRYPTO_DEV) || defined(WOLF_CRYPTO_CB))
 /******************************************************************************/
 /* --- BEGIN wolf Crypto Device Support -- */
 /******************************************************************************/
@@ -3387,7 +3388,7 @@ int wolfTPM2_ClearCryptoDevCb(WOLFTPM2_DEV* dev, int devId)
 /* --- END wolf Crypto Device Support -- */
 /******************************************************************************/
 
-#endif /* WOLF_CRYPTO_DEV */
+#endif /* !WOLFTPM2_NO_WOLFCRYPT && (WOLF_CRYPTO_DEV || WOLF_CRYPTO_CB) */
 
 
 #endif /* !WOLFTPM2_NO_WRAPPER */
